[RESTEASY-1142] ClosedResponseHandlingTest: re-factor layout
