+++ testsuite/integration-tests/src/test/java/org/jboss/resteasy/test/exception/resource/ClosedResponseHandlingPleaseMapExceptionMapper.java
@@ -1,0 +1,18 @@
+package org.jboss.resteasy.test.exception.resource;
+
+import org.jboss.resteasy.test.exception.ClosedResponseHandlingTest;
+
+import javax.ws.rs.core.Response;
+import javax.ws.rs.ext.ExceptionMapper;
+import javax.ws.rs.ext.Provider;
+
+/** ExceptionHandler only uses the logger for tracing exception mapping, so we need to map something.
+ */
+@Provider
+public class ClosedResponseHandlingPleaseMapExceptionMapper implements ExceptionMapper<ClosedResponseHandlingPleaseMapException> {
+
+   @Override
+   public Response toResponse(ClosedResponseHandlingPleaseMapException e) {
+      return e.response;
+   }
+}
